requirements.txt
aws-cdk-lib==2.100.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
Pillow>=10.0.0
requests>=2.28.0

# AWS X-Ray tracing
aws-xray-sdk>=2.12.0

# Structured logging
structlog>=23.0.0
python-json-logger>=2.0.0
